request_counter/requirements.txt
mysql-connector-python==8.0.33
requests==2.28.1
